move block rewards calculation to dedicated spec util
